requirements.txt
Django>=5.0,<6.0
gunicorn==21.2.0
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1